% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{calcTruePrev}
\alias{calcTruePrev}
\title{Function to combine yearly PLWHA prevalence with undiagnosed estimates to provide
true prevalence estimates}
\usage{
calcTruePrev(x, prev)
}
\arguments{
\item{x}{Object of class 'results', the output of combineResults}

\item{prev}{Data frame with 1st column 'Year' and a 2nd column with PLWH
for the population represented in the results}
}
\description{
Function to combine yearly PLWHA prevalence with undiagnosed estimates to provide
true prevalence estimates
}

