% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{compileSubgroups}
\alias{compileSubgroups}
\title{Quickly append results from separate subgroups into one results object}
\usage{
compileSubgroups(x, includeTotal = FALSE)
}
\arguments{
\item{x}{output from runSubgroups}

\item{includeTotal}{Logical indicating whether the 'Total-stratified' estimate
should be included}
}
\value{
Data frame of class 'results'
}
\description{
Takes runSubgroups output and appends results to facilitate plotting primarily
}
\examples{

undx <- runSubgroups(KCsim, subvar='fakeGroup', intLength=1, 
                         cases=c(`Base Case`='base_case'))
all <- compileSubgroups(undx)
}

