\name{estimateProbDist}
\alias{estimateProbDist}
\title{Estimates the discrete time probability distribution function for
the time between infection and diagnosis}
\usage{
  estimateProbDist(infPeriod, intLength = 1)
}
\arguments{
  \item{infPeriod}{a series of times from last HIV test to
  diagnosis}

  \item{intLength}{the length of the discrete time
  intervals}
}
\description{
  Estimates the discrete time probability distribution
  function for the time between infection and diagnosis
}

