% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tid.R
\name{estimateTID}
\alias{estimateTID}
\title{Creates a TID (time from infection to diagnosis) object
 
Calls the TID_pdf() function to estimate the TID from testing history
data for each assumption/case. Evaluates the function for the observed
time span in the data and returns both the probability and cumulative
density distributions.}
\usage{
estimateTID(infPeriod, intLength, cases = NULL)
}
\arguments{
\item{infPeriod}{The vector containing infection periods, or times 
(in years) between last negative test and diagnosis}

\item{intLength}{The interval length by which diagnoses are reported
(also in years, 1=1 year)}

\item{cases}{Cases to estimate; default is c('base_case', 'upper_bound')}
}
\value{
A nested list. The first tier indicates the assumption used
         to estimate the TID. The second tier contains 3 elements:
         "pdffxn", the PDF function, and "pdf" and "cdf" which 
         indicate the respective distributions
}
\description{
Creates a TID (time from infection to diagnosis) object
 
Calls the TID_pdf() function to estimate the TID from testing history
data for each assumption/case. Evaluates the function for the observed
time span in the data and returns both the probability and cumulative
density distributions.
}

