\name{estimateUndiagnosed}
\alias{estimateUndiagnosed}
\title{Estimates the number of undiagnosed in the population}
\usage{
  estimateUndiagnosed(mod, nExt = 500)
}
\arguments{
  \item{mod}{the model}

  \item{nExt}{the number of time units beyond the end of
  the observed period to use in calculation}
}
\description{
  Estimates the number of undiagnosed in the population
}

