% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{estimateUndiagnosedConst}
\alias{estimateUndiagnosedConst}
\title{Estimates the number of undiagnosed in the population, constant incidence}
\usage{
estimateUndiagnosedConst(infPeriod, case, intLength, incidence,
  numSteps = 10000)
}
\arguments{
\item{infPeriod}{A vector of continuous times from last HIV test to diagnosis
for a population}

\item{case}{One of "base_case" or "upper_bound", indicating the
assumption to apply for when infection occurred within infPeriod}

\item{intLength}{A single number indicating the length in years of discrete
time intervals by which HIV diagnoses are recorded. The default of
0.25 represents a quarter-year.}

\item{incidence}{Number of new infections per unit intLength. Assumed
to be constant over time}

\item{numSteps}{Number of discrete time steps used to evaluate the integral}
}
\value{
Scalar representing the number of undiagnosed cases at any
         time point. The precise interval length for this estimate
         will be max(infPeriod)/dt
}
\description{
Uses constant incidence assumption and the TID to estimate undiagnosed
counts per interval
}

