% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{format_eHARS}
\alias{format_eHARS}
\title{Format eHARS person view data for use with the HIVBackCalc package
 
Creates a data frame of class 'testinghistories' with formatted and
calculated variables. Additionally returns a record of formatting
assumptions made for inconsistent entries}
\usage{
format_eHARS(rawdata, assumptionNo = "age16")
}
\arguments{
\item{rawdata}{Unformatted data frame or file path to CSV file}

\item{assumptionNo}{Choice of assumption for those reporting never 
having had a negative test. Default is 'age16', which imputes
time from infection to diagnosis as min(age-16, 18 yrs). 'age16mid'
instead uses the midpoint between age and 16, i.e. 
min((age-16)/2, 18 yrs).}
}
\value{
Data frame of class "testinghistories"
}
\description{
Format eHARS person view data for use with the HIVBackCalc package
 
Creates a data frame of class 'testinghistories' with formatted and
calculated variables. Additionally returns a record of formatting
assumptions made for inconsistent entries
}

