% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{KCplwh}
\alias{KCplwh}
\title{MSM living with HIV in KC 2006-2012}
\format{A data frame with 7 rows and 5 columns:
\describe{
     \item{Year}{Year of estimate}
     \item{White}{MSM living with HIV, White race}
     \item{Black}{MSM living with HIV, Black race}
     \item{Hisp}{MSM living with HIV, Hispanic race}
     \item{Total}{MSM living with HIV}
}}
\source{
Based on estimates from Public Health Seattle King County
}
\usage{
KCplwh
}
\description{
A dataset containing rough estimates of MSM PLWH in King County, by race.
Estimates are specified by year for unit-tests but reflect a time period
average
}
\keyword{datasets}

