% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{KCsim}
\alias{KCsim}
\title{MSM HIV testing histories approximating King County, WA in 2006-2012}
\format{An object of class \code{data.frame} with 1522 rows and 15 columns.}
\source{
Simulation based on data from Public Health Seattle King County
}
\usage{
KCsim
}
\description{
A dataset containing the date of last negative test as well as other characteristics
of HIV diagnoses in a simulated population. The marginal population features approximate
HIV diagnoses among MSM in King County, WA between 2006 and 2012.
}
\keyword{datasets}

