\name{plot.backproj}
\alias{plot.backproj}
\title{plots the model}
\usage{
  plot.backproj(x, time, showDiagCounts = TRUE, ...)
}
\arguments{
  \item{x}{the incedence model}

  \item{time}{a vector of length two giving the start and
  end times of the observed period}

  \item{showDiagCounts}{should diagnosis counts be plotted}

  \item{...}{passed to plot}
}
\description{
  plots the model
}

