% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{plotTestHist}
\alias{plotTestHist}
\title{Plot responses to 'Have you ever had a negative test?' over time}
\usage{
plotTestHist(testhist, panel = NULL)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' with variables
everHadNegTest and yearDx}

\item{panel}{Optional variable by which to panel the plot}
}
\description{
Plots the everHadNegTest results over time
}

