% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.backproj}
\alias{print.backproj}
\title{Print function for "backproj"
 
Print incidence estimates for class "backproj", the output from
estimateIncidence()}
\usage{
\method{print}{backproj}(x, ...)
}
\arguments{
\item{x}{Object of class "backproj"}

\item{...}{passed to cat}
}
\value{
Output to console
}
\description{
Print function for "backproj"
 
Print incidence estimates for class "backproj", the output from
estimateIncidence()
}

