% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Wrapper function that runs an entire analysis}
\usage{
runAnalysis(testhist, descriptives, subvar, intLength, cases,
  runEstimation = TRUE, savedEstimation = FALSE, prev = NULL,
  save = NULL)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' containing the time of 
diagnosis within "timeDx" and time since last negative test in 
"infPeriod"}

\item{descriptives}{NAMED vector with descriptive variables to use in
  looking at the testing histories. See help page for tabTestHist,
  the "variables" argument
subgroups within which to run the backcalculation}

\item{subvar}{Name of the variable within the testhist data frame that defines
subgroups within which to run the backcalculation}

\item{intLength}{Interval length by which diagnoses are tracked; 1=1 year.}

\item{cases}{Vector of names of cases to include for the TID assumptions. 
Defaults to all cases computed by estimateTID, which currently 
offers 'base_case' and 'upper_bound'. Names of vector elements will
be used to label results.}

\item{runEstimation}{Logical indicating whether to run a fresh back-calculation}

\item{savedEstimation}{Logical indicating whether to load a saved back-calculation}

\item{prev}{Optional frame with 1st column 'Year' and a 2nd column with PLWHA
for the population represented in the testhist object}

\item{save}{Optional file path to save compiled results or load saved results}
}
\value{
List object with plot and results list
}
\description{
Ideally this function will successfully produce all the major
plots and results in one go
}
\examples{
# Basic example with a fresh estimation
allRes <- runAnalysis(KCsim, list(c(`Group`='fakeGroup', `Race`='race')),
                     'fakeGroup', 1, c(`Base Case`='base_case',
                                       `Upper Bound`='upper_bound'),
                     runEstimation=TRUE)
# Compute true prevalence and save results
allRes <- runAnalysis(KCsim, list(c(`Group`='fakeGroup', `Race`='race')),
                     'fakeGroup', 1, c(`Base Case`='base_case',
                                       `Upper Bound`='upper_bound'),
                     runEstimation=TRUE,
                     prev=data.frame(`Group 1`=KCplwh$Total/2,
                                     `Group 2`=KCplwh$Total/2,
                                     Total=KCplwh$Total),
                     save='testfile.csv')
# Load the saved estimation and plot true prev results from it
allRes <- runAnalysis(KCsim, list(c(`Group`='fakeGroup', `Race`='race')),
                     'fakeGroup', 1, c(`Base Case`='base_case',
                                       `Upper Bound`='upper_bound'),
                     runEstimation=FALSE,
                     savedEstimation=TRUE,
                     save='testfile.csv')
plotTruePrev(subset(allRes$results$trueprev, 
                   Subgroup=='Total-stratified'))
}

