% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{runBackCalc}
\alias{runBackCalc}
\title{Optional wrapper function to run all the backcalculation steps}
\usage{
runBackCalc(testhist, intLength, cases = NULL, prev = NULL)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' containing the time of 
diagnosis within "timeDx" and time since last negative test in 
"infPeriod"}

\item{intLength}{Interval length by which diagnoses are tracked; 1=1 year.}

\item{cases}{Vector of names of cases to include for the TID assumptions. 
Defaults to all cases computed by estimateTID, which currently 
offers 'base_case' and 'upper_bound'. Names of vector elements will
be used to label results.}

\item{prev}{Optional data frame with 1st column 'Year' and a 2nd column with
PLWHA for the population represented in the testhist object}
}
\description{
Optional wrapper function to run all the backcalculation steps
}

