% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{runSubgroups}
\alias{runSubgroups}
\title{Optional wrapper function to run and compile results for subgroups}
\usage{
runSubgroups(testhist, subvar, intLength, cases = NULL, prev = NULL,
  save = NULL)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' containing the time of 
diagnosis within "timeDx" and time since last negative test in 
"infPeriod"}

\item{subvar}{Name of the variable within the testhist data frame that defines
subgroups within which to run the backcalculation}

\item{intLength}{Interval length by which diagnoses are tracked; 1=1 year.}

\item{cases}{Vector of names of cases to include for the TID assumptions. 
Defaults to all cases computed by estimateTID, which currently 
offers 'base_case' and 'upper_bound'. Names of vector elements will
be used to label results.}

\item{prev}{Optional frame with 1st column 'Year' and a 2nd column with PLWHA
for the population represented in the testhist object}

\item{save}{Optional file path to save compiled true prevalence results}
}
\description{
Optional wrapper function to run and compile results for subgroups
}
\examples{
# Example with fake prevalence corresponding to the fake groups
# Use check.names=FALSE in the prev data frame construction to ensure
# that the column names match the values of the fakeGroup variable
undx <- runSubgroups(KCsim, subvar='fakeGroup', intLength=1, 
                         cases=c(`Base Case`='base_case'),
                         prev=data.frame(Year=KCplwh$Year,
                                         `Group 1`=KCplwh$Total/2,
                                         `Group 2`=KCplwh$Total/2,
                                         Total=KCplwh$Total,
                                         check.names=FALSE))
# Example with saving results
undx <- runSubgroups(KCsim, subvar='fakeGroup', intLength=1, 
                         cases=c(`Base Case`='base_case'),
                         prev=data.frame(Year=KCplwh$Year,
                                         `Group 1`=KCplwh$Total/2,
                                         `Group 2`=KCplwh$Total/2,
                                         Total=KCplwh$Total,
                                         check.names=FALSE),
                         save='testfile.csv')
}

