% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tid.R
\name{summary.TID}
\alias{summary.TID}
\title{Evaluate TID at given time intervals
 
Returns the discrete probability and cumulative density functions
from an object of class 'TID' produced by estimateTID() at specified
time points.}
\usage{
\method{summary}{TID}(x, intLength, times, cases = NULL)
}
\arguments{
\item{x}{The TID object}

\item{intLength}{Diagnosed interval length (from timeDx in the data)
in years}

\item{times}{Vector of times to evaluate, in years. Values must be
multiples of intLength}

\item{cases}{Vector of tidier names for the cases present in the 
TID object, in the same order as names(x). Defaults to
names(x)}
}
\value{
Data frame
}
\description{
Evaluate TID at given time intervals
 
Returns the discrete probability and cumulative density functions
from an object of class 'TID' produced by estimateTID() at specified
time points.
}

