% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{sumResults}
\alias{sumResults}
\title{Summarize results by year}
\usage{
sumResults(x, acrossYears = TRUE, addclass = FALSE)
}
\arguments{
\item{x}{a resultsAll data frame}

\item{acrossYears}{Logical indicating whether to collapse across years}
}
\value{
Original object with resultsSummary or resultsSummaryYear element
}
\description{
Summarize a resultsAll object created by combineResults into 
yearly estimates
}
\examples{

undx <- runBackCalc(KCsim, 1, cases=c(Base='base_case'))
# Compare outputs - they're identical
undx$results$resultsSummary
compare <- sumResults(undx$results, acrossYears=TRUE, addclass=TRUE)
compare$resultsSummary
# Calculate true prev - need to have year strata
compare <- sumResults(undx$results, acrossYears=FALSE, addclass=TRUE)
trueprev <- calcTruePrev(compare, KCplwh[,c('Year', 'Total')])
}

