% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tabTestHist}
\alias{tabTestHist}
\title{Tabulate responses to 'Have you ever had a negative test?'}
\usage{
tabTestHist(testhist, variables, supercolumn = FALSE,
  fullsample_row = FALSE)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' with variable
everHadNegTest}

\item{variables}{NAMED character vector of stratification variable names. 
Names will be used as table labels.
Variables must exist in the testhist data frame.
See examples for using a list structure to do multi-variale cross-tabs.}

\item{supercolumn}{Set to TRUE to include a pretty column indicating
stratification variables}

\item{fullsample_row}{Set to TRUE to have the 1st row be the results
for the entire sample}
}
\description{
Tabulates the everHadNegTest variable by any number of stratification/
subgroup variables
}
\examples{
# One variable
tabTestHist(KCsim, c(Year='yearDx'))
# Marginal, two variables
tabTestHist(KCsim, c(Year='yearDx', Race='race'), supercolumn=TRUE)
# Cross-tab, two variables
tabTestHist(KCsim, list(c(Year='yearDx', Race='race')))
}

