% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tid.R
\name{TID_cdfContinuous}
\alias{TID_cdfContinuous}
\title{Estimates the continuous CDF of time from infection to diagnosis (TID).}
\usage{
TID_cdfContinuous(infPeriod, case, intLength, survivor = FALSE)
}
\arguments{
\item{infPeriod}{A vector containing the continuous time from last
HIV test to diagnosis for each member of the population}

\item{case}{One of "base_case" or "upper_bound", indicating the distributional
assumption used for the hazard of infection within the infPeriod}

\item{intLength}{A single number indicating the length in years of discrete
time intervals by which HIV diagnoses are recorded. The default of
0.25 represents a quarter-year.}

\item{survivor}{Logical indicating whether to return S(x)=1-F(x) instead of
         F(x), the cdf}
}
\value{
A function that takes integer arguments 0 and higher. Will return 0
         for integers greater than floor(max(infPeriod/intLength)) + 1
}
\description{
Uses the window of possible infection (infPeriod) and an assumption
about the distribution of the infection hazard in this interval
to estimate the the CDF of time from infection to diagnosis (TID).
}
\details{
Two infection hazard assumptions are compared in the paper:
a "base case" that assumes the probability of infection is uniformly
distributed in the inter-test interval, and
an "upper bound" that assumes infection occurred immediately after
the last negative test.
}

