% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tid.R
\name{TID_pdf}
\alias{TID_pdf}
\title{Evaluates the discrete PDF of time from infection to diagnosis in reporting intervals.
 
Calls the TID_cdfContinuous function to estimate CDF of the TID, then
evaluates the discrete PDF in time intervals that match reporting period
units (intLength).}
\usage{
TID_pdf(infPeriod, case, intLength)
}
\arguments{
\item{infPeriod}{A vector of continuous times from last HIV test to diagnosis
for a population}

\item{case}{One of "base_case" or "upper_bound", indicating the 
assumption to apply for when infection occurred within infPeriod}

\item{intLength}{A single number indicating the length in years of discrete 
time intervals by which HIV diagnoses are recorded. The default of 
0.25 represents a quarter-year.}
}
\value{
A function that takes integer arguments 0 and higher. Will return 0 
         for integers greater than floor(max(infPeriod/intLength)) + 1
}
\description{
Evaluates the discrete PDF of time from infection to diagnosis in reporting intervals.
 
Calls the TID_cdfContinuous function to estimate CDF of the TID, then
evaluates the discrete PDF in time intervals that match reporting period
units (intLength).
}

