% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aggregateDiagnoses}
\alias{aggregateDiagnoses}
\title{Aggregates diagnosis counts into the specified time interval
 
HIV diagnosis counts from timeDx are aggregated to match 
the specified intLength. Called by tabulateDiagnoses()}
\usage{
aggregateDiagnoses(timeDx, intLength)
}
\arguments{
\item{timeDx}{Time of diagnosis from the testhist data. 
Form should be either YYYY+x/12 where x is the month number, YYYY.25,
YYYY.5, or YYYY.}

\item{intLength}{Desired interval length for diagnoses: 1/12, 
0.25, 0.5 or 1 (1=1 year)}
}
\value{
timeDx variable with intLength
}
\description{
Aggregates diagnosis counts into the specified time interval
 
HIV diagnosis counts from timeDx are aggregated to match 
the specified intLength. Called by tabulateDiagnoses()
}
\seealso{
\code{\link{tabulateDiagnoses}} for example
}

