% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.R
\name{estimateIncidence}
\alias{estimateIncidence}
\title{Backcalculate incidence using HIV diagnoses and TID}
\usage{
estimateIncidence(y, pid, gamma = 0, tol = 10^-5, verbose = FALSE)
}
\arguments{
\item{y}{HIV diagnosis counts vector as formatted by tabulateDiagnoses()}

\item{pid}{Discrete f(x) giving the probability of diagnosis
in the current interval given infection x time units prior, like
  that returned by TID_pdf}

\item{gamma}{Smoothing penalty for EM algorithm}

\item{tol}{The tolerance for the EM algorithm}

\item{verbose}{Logical; print intermediate estimates?}
}
\value{
A list of class "backproj" containing the inputs as well as
         estimated incidence
}
\description{
Uses the standard convolution equation with a customized EM update step
to backcalculate incidence of HIV from diagnoses and the TID aka "pid", time from
infection to diagnosis probability function
}

