% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tid.R
\name{estimateTID}
\alias{estimateTID}
\title{Creates a TID (time from infection to diagnosis) object}
\usage{
estimateTID(infPeriod, intLength)
}
\arguments{
\item{infPeriod}{The vector containing infection periods, or times
(in years) between last negative test and diagnosis}

\item{intLength}{The interval length by which diagnoses are reported
         (also in years, 1=1 year)}
}
\value{
A nested list. The first tier indicates the assumption used
         to estimate the TID. The second tier contains 3 elements:
         "pdffxn", the PDF function, and "pdf" and "cdf" which
         indicate the respective distributions
}
\description{
Calls the TID_pdf() function to estimate the TID from testing history
data for each assumption/case. Evaluates the function for the observed
time span in the data and returns both the probability and cumulative
density distributions.
}

