% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{estimateUndiagnosed}
\alias{estimateUndiagnosed}
\title{Estimates the number of undiagnosed in the population}
\usage{
estimateUndiagnosed(mod, nExt = 500)
}
\arguments{
\item{mod}{Object of class "backproj" containing estimated incidence and TID
function from estimateIncidence()}

\item{nExt}{the number of time units beyond the end of the observed period to use in calculation}
}
\value{
Vector of undiagnosed counts corresponding to time steps in the vector
         created by tabulateDiagnoses() used as input to estimateIncidence()
}
\description{
Uses backcalculated incidence from estimateIncidence() and the TID to
estimate undiagnosed counts per interval
}

