% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{meanEmUpdate}
\alias{meanEmUpdate}
\title{EM update for HIV backcalculation
 
A customized EM update that adjusts for limited HIV surveillance
window and imposes a quadratic smoothing penalty}
\usage{
meanEmUpdate(y, pid, lambda, gamma = 0)
}
\arguments{
\item{y}{HIV diagnosis counts vector as formatted by tabulateDiagnoses()}

\item{pid}{Discrete f(x) giving the probability of diagnosis 
in the current interval given infection x time units prior, like 
  that returned by TID_pdf}

\item{lambda}{Vector of starting values, initialized at average # of
observed diagnoses per time interval}

\item{gamma}{Smoothing penalty for EM algorithm}
}
\value{
Vector of backcalculated incidences for a given step
}
\description{
EM update for HIV backcalculation
 
A customized EM update that adjusts for limited HIV surveillance
window and imposes a quadratic smoothing penalty
}

