% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.backproj}
\alias{plot.backproj}
\title{Plots incidence estimates from estimateIncidence()
 
Plots incidence estimates over time for object of class "backproj" produced 
by estimateIncidence(). Optionally can overlay diagnosis counts.}
\usage{
\method{plot}{backproj}(x, time, showDiagCounts = TRUE, case = "", ...)
}
\arguments{
\item{x}{The "backproj" object containing incidence estimates}

\item{time}{A vector of length two giving the start and end times for the plot.
Allowed to be left unspecified if x$y has names}

\item{showDiagCounts}{Logical: should diagnosis counts be plotted}

\item{...}{Other arguments passed to plot}
}
\value{
Base plot
}
\description{
Plots incidence estimates from estimateIncidence()
 
Plots incidence estimates over time for object of class "backproj" produced 
by estimateIncidence(). Optionally can overlay diagnosis counts.
}

