% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.results}
\alias{plot.results}
\title{Plot estimates incidence and undiagnosed cases
 
Overlays the backcalculated incidence on diagnoses in one panel and
undiagnosed counts in another. Cases are indicated by colors.}
\usage{
\method{plot}{results}(x)
}
\arguments{
\item{x}{List of class "results", the output of combineResults()}
}
\value{
Paneled ggplot2 object
}
\description{
Plot estimates incidence and undiagnosed cases
 
Overlays the backcalculated incidence on diagnoses in one panel and
undiagnosed counts in another. Cases are indicated by colors.
}

