% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{plotDiagnoses}
\alias{plotDiagnoses}
\title{Plots diagnosis counts over time}
\usage{
plotDiagnoses(testhist, panel = NULL)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' with variable
timeDx}

\item{panel}{Name of variable in testhist by which to panel the plot}
}
\description{
Plot of diagnoses vs time, with option to panel by subgroups
}

