% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tabTestHist}
\alias{tabTestHist}
\title{Tabulate responses to 'Have you ever had a negative test?'}
\usage{
tabTestHist(testhist, variables, supercolumn = FALSE,
  fullsample_row = FALSE)
}
\arguments{
\item{testhist}{Data frame of class 'testinghistories' with variable
everHadNegTest}

\item{variables}{Character vector of stratification variable names. 
Variables must exist in the testhist data frame}

\item{supercolumn}{Set to TRUE to include a pretty column indicating
stratification variables}

\item{fullsample_row}{Set to TRUE to have the 1st row be the results
for the entire sample}
}
\description{
Tabulates the everHadNegTest variable by any number of stratification/
subgroup variables
}

