% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tabulateDiagnoses}
\alias{tabulateDiagnoses}
\title{Tabulates HIV diagnoses by interval, as input for estimateIncidence().
 
HIV diagnosis counts for every interval within the calendar time period
of the data are appended to empty intervals that will indicate to 
estimateIncidence() how far back to project incidence counts.}
\usage{
tabulateDiagnoses(testhist, intLength, nPriorInt = 100)
}
\arguments{
\item{testhist}{Object of class 'testinghistories' containing the 
time of diagnosis within "timeDx"}

\item{intLength}{Interval length by which diagnoses are tracked;
1=1 year.}

\item{nPriorInt}{Number of incidence intervals to backproject
prior to earliest observed diagnoses. Use NULL to skip
this and simply return observed diagnosis counts.}
}
\value{
Vector of NA's plus diagnosis counts
}
\description{
Tabulates HIV diagnoses by interval, as input for estimateIncidence().
 
HIV diagnosis counts for every interval within the calendar time period
of the data are appended to empty intervals that will indicate to 
estimateIncidence() how far back to project incidence counts.
}

