% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tid.R
\name{TID_pdf}
\alias{TID_pdf}
\title{Estimates discrete PDF of time from infection to diagnosis.}
\usage{
TID_pdf(infPeriod, case, intLength)
}
\arguments{
\item{infPeriod}{A vector of continuous times from last HIV test to diagnosis
for a population}

\item{case}{One of "base_case" or "upper_bound", indicating the
assumption to apply for when infection occurred within infPeriod}

\item{intLength}{A single number indicating the length in years of discrete
         time intervals by which HIV diagnoses are recorded. The default of
         0.25 represents a quarter-year.}
}
\value{
A function that takes integer arguments 0 and higher. Will return 0
         for integers greater than floor(max(infPeriod/intLength)) + 1
}
\description{
to HIV diagnosis, to estimate a probability distribution of time from HIV
infection to HIV diagnosis (TID). Two "cases" of assumptions are possible:
a "base case" that assumes the probability of infection is uniformly
distributed between the last negative test and diagnosis, and an "upper bound"
that assumes infection occurred instantaneously after the negative test.
}

